stream writes should apply backpressure
Backpressure doesn't make any sense for reads because there's no mechanism to tell IP sources to slow down their packets (although TCP has such a mechanism). For writes, however, the network driver can tell us when it's ready to send the next packet, and block until then. Thus, backpressure on writes actually makes sense.

Here's an example of where write backpressure makes sense. Suppose we want to send out pings to 10000 hosts. We want to go through a loop and send an ICMP packet to each host. Without backpressure, we'd either need a buffer size of >10000, or we'd need to keep re-trying writes until we don't get an error. With backpressure, we can simply go through the loop as usual; writes will block until the packets are ready to fly.